third_party/blink/web_tests/fast/history/state-object-few-arguements-exception-expected.txt

This test makes sure that calls to pushState() and replaceState() with less than 2 arguments fail as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS history.pushState() threw exception TypeError: Failed to execute 'pushState' on 'History': 2 arguments required, but only 0 present..
PASS history.pushState(null) threw exception TypeError: Failed to execute 'pushState' on 'History': 2 arguments required, but only 1 present..
PASS history.pushState(null, null) did not throw exception.
PASS history.pushState(null, null, null) did not throw exception.
PASS history.replaceState() threw exception TypeError: Failed to execute 'replaceState' on 'History': 2 arguments required, but only 0 present..
PASS history.replaceState(null) threw exception TypeError: Failed to execute 'replaceState' on 'History': 2 arguments required, but only 1 present..
PASS history.replaceState(null, null) did not throw exception.
PASS history.replaceState(null, null, null) did not throw exception.
